partner_routes_api/pyproject.toml
[tool.poetry]
name = "partner_routes_api"
version = "0.1.0"
description = ""
authors = ["Bernardo Gomes <bgomesdeabreu@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
Django = "^3.2.7"
djangorestframework = "^3.12.4"
prettyconf = "^2.2.1"
django-simple-history = "^3.0.0"
psycopg2 = "^2.9.1"
kafka-python = "^2.0.2"
factory-boy = "^3.2.0"
django-model-utils = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
